requirements.txt
openai-whisper
torch
sounddevice
numpy
scipy
pyperclip
